Added explicit conversion to double in split's algorithm. Fixes gh-7

+++ LogicElements/src/net/unikernel/bummel/logic_elements/Split/Split.java
@@ -36,8 +36,8 @@ public class Split extends BasicElement
 	@Override
 	public Map<String, Double> process(Map<String, Double> valuesOnPorts)
 	{
-		valuesOnPorts.put(getPorts().get(1), valuesOnPorts.get(getPorts().get(0)));
-		valuesOnPorts.put(getPorts().get(2), valuesOnPorts.get(getPorts().get(0)));
+		valuesOnPorts.put(getPorts().get(1), valuesOnPorts.get(getPorts().get(0)).doubleValue());
+		valuesOnPorts.put(getPorts().get(2), valuesOnPorts.get(getPorts().get(0)).doubleValue());
 		valuesOnPorts.put(getPorts().get(0), 0.);
 		return valuesOnPorts;
 	}
